+++ tests/test_parser_parse_header_lines.py
@@ -2,16 +2,24 @@
 """Test parsing of VCF header lines from strings
 """
 
+import pytest
+
 from vcfpy import header
 from vcfpy import parser
 
 __author__ = 'Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>'
 
+
+@pytest.fixture
+def warning_helper():
+    return parser.WarningHelper()
+
+
 # parser.StupidHeaderLineParser.parse_key_value() --------------------------
 
 
-def test_stupid_vcf_header_line_parser_file_format():
-    p = parser.StupidHeaderLineParser()
+def test_stupid_vcf_header_line_parser_file_format(warning_helper):
+    p = parser.StupidHeaderLineParser(warning_helper)
     INPUT = ('fileFormat', 'VCFv4.2')
     EXPECTED = "HeaderLine('fileFormat', 'VCFv4.2')"
     assert str(p.parse_key_value(*INPUT)) == EXPECTED
@@ -20,8 +28,10 @@ def test_stupid_vcf_header_line_parser_file_format():
 # parser.MappingHeaderLineParser.parse_key_value() -------------------------
 
 
-def test_mapping_vcf_header_line_parser_parse_key_value_filter():
-    p = parser.MappingHeaderLineParser(header.FilterHeaderLine)
+def test_mapping_vcf_header_line_parser_parse_key_value_filter(
+        warning_helper):
+    p = parser.MappingHeaderLineParser(
+        warning_helper, header.FilterHeaderLine)
     INPUT = ('FILTER', '<ID=q10,Description="Quality below 10">')
     EXPECTED = ('FilterHeaderLine(\'FILTER\', \'<ID=q10,Description="'
                 'Quality below 10">\', '
@@ -30,8 +40,10 @@ def test_mapping_vcf_header_line_parser_parse_key_value_filter():
     assert str(p.parse_key_value(*INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_line_parser_parse_key_value_format():
-    p = parser.MappingHeaderLineParser(header.FormatHeaderLine)
+def test_mapping_vcf_header_line_parser_parse_key_value_format(
+        warning_helper):
+    p = parser.MappingHeaderLineParser(
+        warning_helper, header.FormatHeaderLine)
     INPUT = ('FORMAT', '<ID=GT,Number=1,Type=String,Description="Genotype">')
     EXPECTED = ('FormatHeaderLine(\'FORMAT\', \'<ID=GT,Number=1,'
                 'Type=String,Description="Genotype">\', '
@@ -40,8 +52,10 @@ def test_mapping_vcf_header_line_parser_parse_key_value_format():
     assert str(p.parse_key_value(*INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_line_parser_parse_key_value_info():
-    p = parser.MappingHeaderLineParser(header.InfoHeaderLine)
+def test_mapping_vcf_header_line_parser_parse_key_value_info(
+        warning_helper):
+    p = parser.MappingHeaderLineParser(
+        warning_helper, header.InfoHeaderLine)
     INPUT = ('INFO',
              '<ID=NS,Number=1,Type=Integer,Description='
              '"Number of Samples With Data">')
@@ -53,8 +67,10 @@ def test_mapping_vcf_header_line_parser_parse_key_value_info():
     assert str(p.parse_key_value(*INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_line_parser_parse_key_value_contig():
-    p = parser.MappingHeaderLineParser(header.ContigHeaderLine)
+def test_mapping_vcf_header_line_parser_parse_key_value_contig(
+        warning_helper):
+    p = parser.MappingHeaderLineParser(
+        warning_helper, header.ContigHeaderLine)
     INPUT = ('contig',
              '<ID=20,length=62435964,assembly=B36,'
              'md5=f126cdf8a6e0c7f379d618ff66beb2da,'
@@ -71,15 +87,16 @@ def test_mapping_vcf_header_line_parser_parse_key_value_contig():
 
 # parser.HeaderParser.parse_line() -----------------------------------------
 
-def test_vcf_header_parser_file_format():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+
+def test_vcf_header_parser_file_format(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = '##fileFormat=VCFv4.2\n'
     EXPECTED = "HeaderLine('fileFormat', 'VCFv4.2')"
     assert str(p.parse_line(INPUT)) == EXPECTED
 
 
-def test_vcf_header_parser_parse_line_filter():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+def test_vcf_header_parser_parse_line_filter(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = '##FILTER=<ID=q10,Description="Quality below 10">\n'
     EXPECTED = ('FilterHeaderLine(\'FILTER\', \'<ID=q10,Description="'
                 'Quality below 10">\', '
@@ -88,8 +105,8 @@ def test_vcf_header_parser_parse_line_filter():
     assert str(p.parse_line(INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_parser_parse_line_format():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+def test_mapping_vcf_header_parser_parse_line_format(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = '##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">\n'
     EXPECTED = ('FormatHeaderLine(\'FORMAT\', \'<ID=GT,Number=1,'
                 'Type=String,Description="Genotype">\', '
@@ -98,8 +115,8 @@ def test_mapping_vcf_header_parser_parse_line_format():
     assert str(p.parse_line(INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_parser_parse_line_info():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+def test_mapping_vcf_header_parser_parse_line_info(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = ('##INFO='
              '<ID=NS,Number=1,Type=Integer,Description='
              '"Number of Samples With Data">\n')
@@ -111,8 +128,8 @@ def test_mapping_vcf_header_parser_parse_line_info():
     assert str(p.parse_line(INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_parser_parse_line_contig():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+def test_mapping_vcf_header_parser_parse_line_contig(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = ('##contig='
              '<ID=20,length=62435964,assembly=B36,'
              'md5=f126cdf8a6e0c7f379d618ff66beb2da,'
@@ -128,8 +145,8 @@ def test_mapping_vcf_header_parser_parse_line_contig():
     assert str(p.parse_line(INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_parser_parse_line_alt_allele():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+def test_mapping_vcf_header_parser_parse_line_alt_allele(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = ('##ALT='
              '<ID=R,Description="IUPAC code R = A/G">\n')
     EXPECTED = ("AltAlleleHeaderLine('ALT', "
@@ -140,8 +157,8 @@ def test_mapping_vcf_header_parser_parse_line_alt_allele():
     assert str(p.parse_line(INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_parser_parse_line_meta():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+def test_mapping_vcf_header_parser_parse_line_meta(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = ('##META='
              '<ID=Assay,Type=String,Number=.,Values=[WholeGenome, Exome]>\n')
     EXPECTED = (
@@ -153,8 +170,8 @@ def test_mapping_vcf_header_parser_parse_line_meta():
     assert str(p.parse_line(INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_parser_parse_line_pedigree():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+def test_mapping_vcf_header_parser_parse_line_pedigree(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = ('##PEDIGREE='
              '<ID=TumourSample,Original=GermlineID>\n')
     EXPECTED = ("PedigreeHeaderLine('PEDIGREE', "
@@ -165,8 +182,8 @@ def test_mapping_vcf_header_parser_parse_line_pedigree():
     assert str(p.parse_line(INPUT)) == EXPECTED
 
 
-def test_mapping_vcf_header_parser_parse_line_sample():
-    p = parser.HeaderParser(parser.HEADER_PARSERS)
+def test_mapping_vcf_header_parser_parse_line_sample(warning_helper):
+    p = parser.HeaderParser(parser.build_header_parsers(warning_helper))
     INPUT = ('##SAMPLE='
              '<ID=Sample1,Assay=WholeGenome,Ethnicity=AFR,Disease=None,'
              'Description="Patient germline genome from unaffected",'