+++ vcfpy/header.py
@@ -8,6 +8,7 @@ import json
 import sys
 
 from . import exceptions
+from .warn_utils import WarningHelper
 
 try:
     from cyordereddict import OrderedDict
@@ -40,9 +41,156 @@ HEADER_NUMBER_GENOTYPES = 'G'
 HEADER_NUMBER_UNBOUNDED = '.'
 
 
-def _warn(msg):
-    """Print warning message in case of missing attributes"""
-    print('[vcfpy] WARNING: {}'.format(msg), file=sys.stderr)
+class FieldInfo:
+    """Core information for describing field type and number"""
+
+    def __init__(self, type_, number, description=None):
+        #: The type, one of INFO_TYPES or FORMAT_TYPES
+        self.type = type_
+        #: Number description, either an int or constant
+        self.number = number
+        #: Description for the header field, optional
+        self.description = description
+
+    def __str__(self):
+        return 'FieldInfo({}, {}, {})'.format(
+            *map(repr, [self.type, self.number, self.description]))
+
+    def __repr__(self):
+        return str(self)
+
+
+# Reserved INFO keys ----------------------------------------------------------
+
+#: Reserved fields for INFO from VCF v4.3
+RESERVED_INFO = {
+    # VCF v4.3, Section 1.6.1
+    'AA': FieldInfo('String', 1, 'Ancestral Allele'),
+    'AC': FieldInfo('Integer', 'A',
+                    'Allele count in genotypes, for each ALT allele, in the '
+                    'same order as listed'),
+    'AD': FieldInfo('Integer', 'R', 'Total read depth for each allele'),
+    'ADF': FieldInfo('Integer', 'R', 'Forward read depth for each allele'),
+    'ADR': FieldInfo('Integer', 'R', 'Reverse read depth for each allele'),
+    'AF': FieldInfo('Float', 'A',
+                    'Allele frequency for each ALT allele in the same order '
+                    'as listed: used for estimating from primary data not '
+                    'called genotypes'),
+    'AN': FieldInfo('Integer', 1,
+                    'Total number of alleles in called genotypes'),
+    'BQ': FieldInfo('Float', 1, 'RMS base quality at this position'),
+    'CIGAR': FieldInfo('String', 'A',
+                       'CIGAR string describing how to align each ALT allele '
+                       'to the reference allele'),
+    'DB': FieldInfo('Flag', 0, 'dbSNP membership'),
+    'DP': FieldInfo('Integer', 1, 'Combined depth across samples'),
+    'H2': FieldInfo('Flag', 0, 'Membership in HapMap 2'),
+    'H3': FieldInfo('Flag', 0, 'Membership in HapMap 3'),
+    'MQ': FieldInfo('Integer', 1, 'RMS mapping quality'),
+    'MQ0': FieldInfo('Integer', 1,
+                     'Number of MAPQ == 0 reads covering this record'),
+    'NS': FieldInfo('Integer', 1, 'Number of samples with data'),
+    'SB': FieldInfo('Integer', 4, 'Strand bias at this position'),
+    'SOMATIC': FieldInfo('Flag', 0,
+                         'Indicates that the record is a somatic mutation, '
+                         'for cancer genomics'),
+    'VALIDATED': FieldInfo('Flag', 0, 'Validated by follow-up experiment'),
+    '1000G': FieldInfo('Flag', 0, 'Membership in 1000 Genomes'),
+
+    # VCF v4.3, Section 3
+    'IMPRECISE': FieldInfo('Flag', 0, 'Imprecise structural variation'),
+    'NOVEL': FieldInfo('Flag', 0, 'Indicates a novel structural variation'),
+    'END': FieldInfo('Integer', 1,
+                     'End position of the variant described in this record '
+                     '(for symbolic alleles)'),
+    'SVTYPE': FieldInfo('String', 1, 'Type of structural variant'),
+    'SVLEN': FieldInfo('Integer', 1,
+                       'Difference in length between REF and ALT alleles'),
+    'CIPOS': FieldInfo('Integer', 2,
+                       'Confidence interval around POS for imprecise '
+                       'variants'),
+    'CIEND': FieldInfo('Integer', 2,
+                       'Confidence interval around END for imprecise '
+                       'variants'),
+    'HOMLEN': FieldInfo('Integer', '.',
+                        'Length of base pair identical micro-homology at '
+                        'event breakpoints'),
+    'HOMSEQ': FieldInfo('String', '.',
+                        'Sequence of base pair identical micro-homology at '
+                        'event breakpoints'),
+    'BKPTID': FieldInfo('String', '.',
+                        'ID of the assembled alternate allele in the '
+                        'assembly file'),
+    'MEINFO': FieldInfo('String', 4,
+                        'Mobile element info of the form '
+                        'NAME,START,END,POLARITY'),
+    'METRANS': FieldInfo('String', 4,
+                         'Mobile element transduction info of the form '
+                         'CHR,START,END,POLARITY'),
+    'DGVID': FieldInfo('String', 1,
+                       'ID of this element in Database of Genomic Variation'),
+    'DBVARID': FieldInfo('String', 1, 'ID of this element in DBVAR'),
+    'DBRIPID': FieldInfo('String', 1, 'ID of this element in DBRIP'),
+    'MATEID': FieldInfo('String', '.', 'ID of mate breakends'),
+    'PARID': FieldInfo('String', 1, 'ID of partner breakend'),
+    'EVENT': FieldInfo('String', 1, 'ID of event associated to breakend'),
+    'CILEN': FieldInfo('Integer', 2,
+                       'Confidence interval around the inserted material '
+                       'between breakends'),
+    'DP': FieldInfo('Integer', 1, 'Read Depth of segment containing breakend'),
+    'DPADJ': FieldInfo('Integer', '.', 'Read Depth of adjacency'),
+    'CN': FieldInfo('Integer', 1,
+                    'Copy number of segment containing breakend'),
+    'CNADJ': FieldInfo('Integer', '.', 'Copy number of adjacency'),
+    'CICN': FieldInfo('Integer', 2,
+                      'Confidence interval around copy number for the '
+                      'segment'),
+    'CICNADJ': FieldInfo('Integer', '.',
+                         'Confidence interval around copy number for the '
+                         'adjacency'),
+}
+
+# Reserved FORMAT keys --------------------------------------------------------
+
+RESERVED_FORMAT = {
+    # VCF v 4.3, Section 1.6.2
+    'AD': FieldInfo('Integer', 'R', 'Total, per-sample read depth'),
+    'ADF': FieldInfo('Integer', 'R', 'Forward-strand, per-sample read depth'),
+    'ADR': FieldInfo('Integer', 'R', 'Reverse-strand, per-sample read depth'),
+    'DP': FieldInfo('Integer', 1,
+                    'Read depth at this position for this sample'),
+    'EC': FieldInfo('Integer', 'A',
+                    'Expected alternate allele counts for each alternate '
+                    'allele'),
+    'FT': FieldInfo('String', '.', 'Filters applied for this sample'),
+    'GQ': FieldInfo('Integer', 'G',
+                    'Phred-scale, conditional genotype quality'),
+    'GP': FieldInfo('Float', 'G', 'Genotype posterior probabilities'),
+    'GT': FieldInfo('String', 1, 'Genotype call'),
+    'GL': FieldInfo('Float', 'G', 'Log10-scaled likelihoods for genotypes'),
+    'HQ': FieldInfo('Integer', 2, 'Haplotype qualities'),
+    'MQ': FieldInfo('Integer', 1, 'RMS mapping quality'),
+    'PL': FieldInfo('Integer', 'G',
+                    'Phred-scaled genotype likelihoods, rounded to integers'),
+    'PQ': FieldInfo('Integer', 1, 'Phasing quality'),
+    'PS': FieldInfo('Integer', 1,
+                    'Non-negative 32 bit integer giving phasing set '
+                    'for this sample and this chromosome'),
+
+    # VCF v4.3, Section 4
+    'CN': FieldInfo('Integer', 1, 'Copy number genotype for imprecise events'),
+    'CNQ': FieldInfo('Float', 1,
+                     'Copy number genotype quality for imprecise events'),
+    'CNL': FieldInfo('Float', 'G',
+                     'Copy number genotype likelihood for imprecise events'),
+    'CNP': FieldInfo('Float', 'G',
+                     'Copy number posterior probabilities'),
+    'NQ': FieldInfo('Integer', 1,
+                    'Phred style probability score that the variant is novel'),
+    'HAP': FieldInfo('Integer', 1, 'Unique haplotype identifier'),
+    'AHAP': FieldInfo('Integer', 1,
+                      'Unique identifier of ancestral haplotype'),
+}
 
 
 # header files to enforce double-quoting for
@@ -62,22 +210,6 @@ def serialize_for_header(key, value):
         return '[{}]'.format(', '.join(value))
     else:
         return str(value)
-
-
-class FieldInfo:
-    """Core information for describing field type and number"""
-
-    def __init__(self, type_, number):
-        #: The type, one of INFO_TYPES or FORMAT_TYPES
-        self.type = type_
-        #: Number description, either an int or constant
-        self.number = number
-
-    def __str__(self):
-        return 'FieldInfo({}, {})'.format(*map(repr, [self.type, self.number]))
-
-    def __repr__(self):
-        return str(self)
 
 
 def header_without_lines(header, remove):
@@ -113,11 +245,13 @@ class Header:
     a filtered list of header lines.
     """
 
-    def __init__(self, lines=[], samples=None):
+    def __init__(self, lines=[], samples=None, warning_helper=WarningHelper()):
         #: ``list`` of :py:HeaderLine objects
         self.lines = list(lines)
         #: :py:class:`SamplesInfo` object
         self.samples = samples
+        # helper for printing warnings
+        self.warning_helper = warning_helper
         # build indices for the different field types
         self._indices = self._build_indices()
 
@@ -128,8 +262,9 @@ class Header:
             if line.key in LINES_WITH_ID:
                 result.setdefault(line.key, OrderedDict())
                 if line.mapping['ID'] in result[line.key]:
-                    _warn(('Seen {} header more than once: {}, using first'
-                           'occurence').format(line.key, line.mapping['ID']))
+                    self.warning_helper.warn_once(
+                        ('Seen {} header more than once: {}, using first'
+                         'occurence').format(line.key, line.mapping['ID']))
                 else:
                     result[line.key][line.mapping['ID']] = line
             else:
@@ -179,9 +314,10 @@ class Header:
         if not hasattr(header_line, 'mapping'):
             return  # no registration required
         if header_line.mapping['ID'] in self._indices[header_line.key]:
-            _warn(('Detected duplicate header line with type {} and ID {}. '
-                   'Ignoring this and subsequent one').format(
-                       header_line.key, header_line.mapping['ID']))
+            self.warning_helper.warn_once(
+                ('Detected duplicate header line with type {} and ID {}. '
+                 'Ignoring this and subsequent one').format(
+                     header_line.key, header_line.mapping['ID']))
         else:
             self._indices[header_line.key][
                 header_line.mapping['ID']] = header_line
@@ -198,9 +334,15 @@ class Header:
         result = self._indices[type_].get(key)
         if result:
             return result
-        _warn('{} {} not found using String/"." instead'.format(
-            type_, key))
-        return FieldInfo('String', HEADER_NUMBER_UNBOUNDED)
+        if key in RESERVED_INFO:
+            res = FieldInfo(RESERVED_INFO[key].type,
+                            RESERVED_INFO[key].number)
+        else:
+            res = FieldInfo('String', HEADER_NUMBER_UNBOUNDED)
+        self.warning_helper.warn_once(
+            '{} {} not found using {}/{} instead'.format(
+                type_, key, res.type, repr(res.number)))
+        return res
 
     def __str__(self):
         tpl = 'Header(lines={}, samples={})'
@@ -214,11 +356,13 @@ class HeaderLine:
     """Base class for VCF header lines
     """
 
-    def __init__(self, key, value):
+    def __init__(self, key, value, warning_helper=WarningHelper()):
         #: ``str`` with key of header line
         self.key = key
         # ``str`` with raw value of header line
         self._value = value
+        #: Helper for printing warnings
+        self.warning_helper = warning_helper
 
     @property
     def value(self):
@@ -255,8 +399,8 @@ class SimpleHeaderLine(HeaderLine):
         the case of missing key ``"ID"``
     """
 
-    def __init__(self, key, value, mapping):
-        super().__init__(key, value)
+    def __init__(self, key, value, mapping, warning_helper=WarningHelper()):
+        super().__init__(key, value, warning_helper)
         # check existence of key "ID"
         if 'ID' not in mapping:
             raise exceptions.InvalidHeaderException(
@@ -289,7 +433,7 @@ class AltAlleleHeaderLine(SimpleHeaderLine):
         """Construct from mapping, not requiring the string value"""
         return AltAlleleHeaderLine('ALT', mapping_to_str(mapping), mapping)
 
-    def __init__(self, key, value, mapping):
+    def __init__(self, key, value, mapping, warning_helper=WarningHelper()):
         super().__init__(key, value, mapping)
         #: name of the alternative allele
         self.id = self.mapping['ID']
@@ -310,13 +454,13 @@ class ContigHeaderLine(SimpleHeaderLine):
         """Construct from mapping, not requiring the string value"""
         return ContigHeaderLine('contig', mapping_to_str(mapping), mapping)
 
-    def __init__(self, key, value, mapping):
-        super().__init__(key, value, mapping)
+    def __init__(self, key, value, mapping, warning_helper=WarningHelper()):
+        super().__init__(key, value, mapping, warning_helper)
         # convert 'length' entry to integer if possible
         if 'length' in self.mapping:
             mapping['length'] = int(mapping['length'])
         else:
-            _warn(
+            self.warning_helper.warn_once(
                 'Field "length" not found in header line {}={}'.format(
                     key, value))
         #: name of the contig
@@ -338,11 +482,11 @@ class FilterHeaderLine(SimpleHeaderLine):
         """Construct from mapping, not requiring the string value"""
         return FilterHeaderLine('FILTER', mapping_to_str(mapping), mapping)
 
-    def __init__(self, key, value, mapping):
-        super().__init__(key, value, mapping)
+    def __init__(self, key, value, mapping, warning_helper=WarningHelper()):
+        super().__init__(key, value, mapping, warning_helper)
         # check for "Description" key
         if 'Description' not in self.mapping:
-            _warn(
+            self.warning_helper.warn_once(
                 'Field "Description" not found in header line {}={}'.format(
                     key, value))
         #: token for the filter
@@ -366,7 +510,8 @@ class MetaHeaderLine(SimpleHeaderLine):
         """Construct from mapping, not requiring the string value"""
         return MetaHeaderLine('META', mapping_to_str(mapping), mapping)
 
-    def __init__(self, key, value, mapping):
+    def __init__(self, key, value, mapping,
+                 warning_helper=WarningHelper()):
         super().__init__(key, value, mapping)
         #: name of the alternative allele
         self.id = self.mapping['ID']
@@ -385,7 +530,8 @@ class PedigreeHeaderLine(SimpleHeaderLine):
         """Construct from mapping, not requiring the string value"""
         return PedigreeHeaderLine('PEDIGREE', mapping_to_str(mapping), mapping)
 
-    def __init__(self, key, value, mapping):
+    def __init__(self, key, value, mapping,
+                 warning_helper=WarningHelper()):
         super().__init__(key, value, mapping)
         #: name of the alternative allele
         self.id = self.mapping['ID']
@@ -404,7 +550,8 @@ class SampleHeaderLine(SimpleHeaderLine):
         """Construct from mapping, not requiring the string value"""
         return PedigreeHeaderLine('SAMPLE', mapping_to_str(mapping), mapping)
 
-    def __init__(self, key, value, mapping):
+    def __init__(self, key, value, mapping,
+                 warning_helper=WarningHelper()):
         super().__init__(key, value, mapping)
         #: name of the alternative allele
         self.id = self.mapping['ID']
@@ -420,7 +567,8 @@ class CompoundHeaderLine(HeaderLine):
     Compound header lines describe fields that can have more than one entry.
     """
 
-    def __init__(self, key, value, mapping):
+    def __init__(self, key, value, mapping,
+                 warning_helper=WarningHelper()):
         super().__init__(key, value)
         #: OrderedDict with key/value mapping
         self.mapping = OrderedDict(mapping.items())
@@ -476,7 +624,8 @@ class InfoHeaderLine(CompoundHeaderLine):
         """Construct from mapping, not requiring the string value"""
         return InfoHeaderLine('INFO', mapping_to_str(mapping), mapping)
 
-    def __init__(self, key, value, mapping):
+    def __init__(self, key, value, mapping,
+                 warning_helper=WarningHelper()):
         super().__init__(key, value, mapping)
         #: key in the INFO field
         self.id = self.mapping['ID']
@@ -485,12 +634,12 @@ class InfoHeaderLine(CompoundHeaderLine):
         # check for "Type" field
         type_ = self.mapping.get('Type')
         if 'Type' not in self.mapping:
-            _warn(
+            self.warning_helper.warn_once(
                 ('Field "Type" not found in header line, using String '
                  'instead {}={}').format(key, value))
             type_ = 'String'
         if 'Type' in self.mapping and type_ not in INFO_TYPES:
-            _warn(
+            self.warning_helper.warn_once(
                 ('Invalid INFO value type {} in header line, using String '
                  'instead, {}={}').format(self.mapping['Type'], key, value))
             type_ = 'String'
@@ -498,7 +647,7 @@ class InfoHeaderLine(CompoundHeaderLine):
         self.type = type_
         # check for "Description" key
         if 'Description' not in self.mapping:
-            _warn(
+            self.warning_helper.warn_once(
                 'Field "Description" not found in header line {}={}'.format(
                     key, value))
         #: description, should be given, ``None`` if not given
@@ -522,7 +671,8 @@ class FormatHeaderLine(CompoundHeaderLine):
         """Construct from mapping, not requiring the string value"""
         return FormatHeaderLine('FORMAT', mapping_to_str(mapping), mapping)
 
-    def __init__(self, key, value, mapping):
+    def __init__(self, key, value, mapping,
+                 warning_helper=WarningHelper()):
         super().__init__(key, value, mapping)
         #: key in the INFO field
         self.id = self.mapping['ID']
@@ -531,12 +681,12 @@ class FormatHeaderLine(CompoundHeaderLine):
         # check for "Type" field
         type_ = self.mapping.get('Type')
         if 'Type' not in self.mapping:
-            _warn(
+            self.warning_helper.warn_once(
                 ('Field "Type" not found in header line, using String '
                  'instead {}={}').format(key, value))
             type_ = 'String'
         if 'Type' in self.mapping and type_ not in FORMAT_TYPES:
-            _warn(
+            self.warning_helper.warn_once(
                 ('Invalid INFO value type {} in header line, using String '
                  'instead, {}={}').format(self.mapping['Type'], key, value))
             type_ = 'String'
@@ -544,7 +694,7 @@ class FormatHeaderLine(CompoundHeaderLine):
         self.type = type_
         # check for "Description" key
         if 'Description' not in self.mapping:
-            _warn(
+            self.warning_helper.warn_once(
                 'Field "Description" not found in header line {}={}'.format(
                     key, value))
         #: description, should be given, ``None`` if not given