+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.5.0 (2016-09-24)
+------------------
+
+* Deactivating warnings on record parsing by default because of performance
+* Adding validation for ``INFO`` and ``FORMAT`` fields on reading (#8)
+* Adding predefined ``INFO`` and ``FORMAT`` fields to ``pyvcf.header`` (#32)
+
 0.4.1 (2016-09-22)
 ------------------